Styling Index
I am actually quite impressed that I managed to pull this off, but I guess that's problem with my confidence level.

Anyway, I haven't finished adding bits and styling index so I'm going to do that next, but I managed to add and style the hero bit (I thinkt that's what it's called) and the steps bit, according to the design.