Search System Finished
I have finished working on the search system as for now. You can now search books (for now) by slowly typing in the name of what you are looking for. You can also tailor down the search by ticking certain tags, which took me a while to get working but it works.

I was smashing my head about the invalid parameters error, and it turned out that I just had the prepare command in the wrong place.